:recyle: Extract to components
